pyproject.toml
[project]
name = "crew"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "crewai",
    "requests",
    "python-dotenv",
    "pandas",
    "pyyaml",
    "streamlit>=1.46.1",
    "crewai-tools>=0.49.0",
    "load-dotenv>=0.1.0",
    "plotly>=6.2.0",
]
